Convert line endings to LF (some had only CR) (read commit desc.).
Carriage returns (CR) get converted to \r with Javascript, which is ignored in HTML. Line feeds (LF) are converted to \n by Javascript, which is used as newline in HTML. Some text comes with only CR on some lines, thus the line endings are ignored. This is a fix until the model fixes its results when sending to the view.